requirements.txt
pytest
mcp[cli]
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
